(maint) Use Bolt::PAL for interactions with puppet
Previously the PAL class was only used for tests and was duplicated in
CLI. This converts CLI to interact with Puppet through the PAL clas
instead.
