+++ spec/bolt/pal_spec.rb
@@ -1,0 +1,2 @@
+# Currently PAL is tested through the CLI execute method. New tests should be
+# added here as we add features or fix bugs.
