+++ spec/spec_helper.rb
@@ -4,6 +4,8 @@ require 'bolt'
 require 'bolt/logger'
 require 'logging'
 require 'rspec/logging_helper'
+# Make sure puppet is required for the 'reset puppet settings' context
+require_relative '../vendored/require_vendored'
 
 $LOAD_PATH.unshift File.join(__dir__, 'lib')
